calendar-booking-agent/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
langchain==0.1.0
langchain-google-genai==0.0.10
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
python-multipart==0.0.6